+++ NEWS.md
@@ -1,5 +1,11 @@
 # assignR news
 
+## assignR 2.2.0
+* Add wDist function and c and plot methods for summarizing weighted distance and bearing distributions using sample collection locations and posterior probability maps
+* QA option to run without iterative recalibration of isoscape
+* Bug fixes
+* Documentation edits
+
 ## assignR 2.1.1
 * Bug fixes